Added libnoise library and used perlin noise to randomise fuel load; Moved camera clamping to after movement to prevent flickering effect; Added a number of new SVG files

+++ Assets/Scripts/TerrainData.cs
@@ -12,7 +12,7 @@ public class TerrainData : MonoBehaviour
         FOREST,
     };
 
-    public int fuelLoad;
+    public float fuelLoad;
     public int fireLevel;
     public int tempFireLevel;
